refactor: common 컴포넌트 style 파일 분리
