+++ src/components/@common/ErrorBoundary/index.jsx
@@ -1,7 +1,5 @@
 import { Component } from 'react';
-import styled from 'styled-components';
-import { flexCenter } from 'styles/mixin';
-import theme from 'styles/theme';
+import { Styled } from './styles';
 
 export default class ErrorBoundary extends Component {
   constructor(props) {
@@ -21,11 +19,3 @@ export default class ErrorBoundary extends Component {
     return this.props.children;
   }
 }
-
-const Styled = {
-  Message: styled.div`
-    ${flexCenter}
-    font-size: 5rem;
-    color: ${theme.colors.grey};
-  `,
-};
