+++ src/components/@common/Loading/index.tsx
@@ -1,0 +1,14 @@
+import { Styled } from './styles';
+
+const Loading = () => {
+  return (
+    <Styled.Loader>
+      <Styled.Spinner>
+        <Styled.Debounce1 />
+        <Styled.Debounce2 />
+      </Styled.Spinner>
+    </Styled.Loader>
+  );
+};
+
+export default Loading;
